backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
sqlalchemy
alembic
python-dotenv
openpyxl
python-dateutil
python-multipart